Handling additional keys
Allowing the editor to read and process the home, end, page up and page down keys. For now, page up takes the cursor to the begining of the terminal, page down takes it to the end, home takes it to the far left and end takes it to the far right. Also the delete key is being read but not processed